requirements.txt
Flask
numpy
scipy
sounddevice
openai-whisper
pyttsx3
google-generativeai
